requirements.txt
asgiref==3.8.1
Django==5.0.4
greenlet==3.0.3
playwright==1.43.0
pyee==11.1.0
ruff==0.4.1
sqlparse==0.5.0
typing_extensions==4.11.0
psycopg2-binary==2.9.9
pillow==10.3.0
azure-storage-blob==12.20.0